backend/requirements.txt
langchain>=0.1.0
langchain-core>=0.1.9
langchain-openai>=0.0.5
openai>=1.3.0
fastapi>=0.104.1
uvicorn>=0.24.0
sqlalchemy>=2.0.23
python-dotenv>=1.0.0
pydantic>=2.0.0
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
pydantic-settings==2.1.0
pytest==7.4.0
httpx==0.25.2
psycopg2-binary>=2.9.5
email-validator>=2.0.0
bcrypt>=4.0.1
starlette>=0.27.0
aiohttp>=3.8.1
websocket-client>=1.7.0 